backend/requirements.txt
alembic==1.11.1
fastapi==0.110.1
httpx==0.27.0
pyjwt==2.8.0
itsdangerous==2.1.2
sqladmin==0.17.0
sqlalchemy==2.0.29
sqlalchemy-utils==0.41.2
asyncpg==0.29.0
passlib==1.7.4
pydantic-settings==2.2.1
uvicorn==0.29.0
requests==2.32.3
websockets==13.0

beautifulsoup4==4.12.3
chromadb==0.4.24
mattsollamatools==0.0.21
ollama==0.1.8
python_magic==0.4.27
numpy==1.26.4
pypdf==4.3.1